Se agrega la acción de grenar en el automovil estandar y se agrega una clase Main

+++ VehiculoTransporte/VehiculoTransporte.java
@@ -21,7 +21,7 @@ public abstract class VehiculoTransporte{
         return velocidad;
     }
 
-    public double modoFrenar(){
+    public float modoFrenar(){
         return frenar.frenar(intensidad);
     }
     public float modoAcelerar(){
